backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==2.1.1
# yfinance removed - using direct Yahoo Finance REST API for stability
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
sse-starlette==2.1.3
scipy==1.14.1
